'Use of Global Variables section' issue
Under 'Use of global variables' section, I tried following the approach suggested, but I could not get it to work. I found 'Player.players' had to be added to the 'add_player' method for this method could work. This is so that we create the empty hash to which we can then add players.